# docs/swagger.yaml
swagger: "2.0"
info:
  title: "E-commerce API"
  description: "API para gestión de pedidos, productos y notificaciones en un sistema de e-commerce"
  version: "1.0.0"
  contact:
    name: "UTN DSW Grupo 05"
host: "localhost:3000"
basePath: "/"
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /pedidos:
    post:
      summary: "Crear nuevo pedido"
      description: "Crea un nuevo pedido en el sistema"
      operationId: "crearPedido"
      tags:
        - Pedidos
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/PedidoCreate"
      responses:
        "201":
          description: "Pedido creado exitosamente"
          schema:
            $ref: "#/definitions/Pedido"
        "400":
          description: "Datos de entrada inválidos"
          schema:
            $ref: "#/definitions/Error"
  /pedidos/{pedido_id}/cancelacion:
    put:
      summary: "Cancelar pedido"
      description: "Cancela un pedido existente"
      operationId: "cancelarPedido"
      tags:
        - Pedidos
      parameters:
        - name: pedido_id
          in: path
          required: true
          type: string
          description: "ID del pedido"
      responses:
        "200":
          description: "Pedido cancelado exitosamente"
          schema:
            $ref: "#/definitions/Pedido"
        "404":
          description: "Pedido no encontrado"
          schema:
            $ref: "#/definitions/Error"
        "400":
          description: "El pedido no puede ser cancelado"
          schema:
            $ref: "#/definitions/Error"
  /pedidos/usuario/{usuario_id}:
    get:
      summary: "Obtener pedidos por usuario"
      description: "Obtiene todos los pedidos de un usuario específico"
      operationId: "obtenerPedidos"
      tags:
        - Pedidos
      parameters:
        - name: usuario_id
          in: path
          required: true
          type: string
          description: "ID del usuario"
      responses:
        "200":
          description: "Lista de pedidos del usuario"
          schema:
            type: array
            items:
              $ref: "#/definitions/Pedido"
        "404":
          description: "Usuario no encontrado"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Pedido:
    type: object
    properties:
      _id:
        type: string
      comprador:
        type: string
      items:
        type: array
        items:
          $ref: "#/definitions/ItemPedido"
      moneda:
        type: string
        enum: [PESO_ARG, DOLAR_USA, REAL]
      total:
        type: number
      direccionEntrega:
        $ref: "#/definitions/DireccionEntrega"
      estado:
        type: string
        enum:
          [PENDIENTE, CONFIRMADO, EN_PREPARACION, ENVIADO, ENTREGADO, CANCELADO]
      fechaCreacion:
        type: string
        format: date-time
      historialEstados:
        type: array
        items:
          $ref: "#/definitions/CambioEstado"
  PedidoCreate:
    type: object
    required: [comprador, items, moneda, direccionEntrega]
    properties:
      comprador:
        type: string
      items:
        type: array
        items:
          $ref: "#/definitions/ItemPedidoCreate"
      moneda:
        type: string
        enum: [PESO_ARG, DOLAR_USA, REAL]
      direccionEntrega:
        $ref: "#/definitions/DireccionEntrega"
  ItemPedido:
    type: object
    properties:
      producto:
        type: string
      cantidad:
        type: integer
      precio:
        type: number
      subtotal:
        type: number
  ItemPedidoCreate:
    type: object
    required: [producto, cantidad]
    properties:
      producto:
        type: string
      cantidad:
        type: integer
  DireccionEntrega:
    type: object
    properties:
      calle:
        type: string
      numero:
        type: string
      ciudad:
        type: string
      codigoPostal:
        type: string
      provincia:
        type: string
      pais:
        type: string
  CambioEstado:
    type: object
    properties:
      fecha:
        type: string
        format: date-time
      estado:
        type: string
        enum:
          [PENDIENTE, CONFIRMADO, EN_PREPARACION, ENVIADO, ENTREGADO, CANCELADO]
      usuario:
        type: string
      motivo:
        type: string
...
  Error:
    type: object
    properties:
      success:
        type: boolean
      error:
        type: string